refactor: moved Calendar component from Apartment/Create to components
